@@ -3,7 +3,7 @@ description: Demonstrates how to use the social_share plugin.
 publish_to: "none"
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -11,11 +11,11 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  screenshot: ^0.1.1
-  path_provider: ^1.5.1
-  http:
-  image_picker: ^0.6.3+1
+  cupertino_icons: ^1.0.3
+  screenshot: ^1.2.1
+  path_provider: ^2.0.2
+  http: ^0.13.3
+  image_picker: ^0.8.1+3
 
 dev_dependencies:
   flutter_test:
