@@ -5,13 +5,13 @@ author: <sskshekar10@gmail.com>
 homepage: https://github.com/ShekarMudaliyar/social_share
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  path_provider: ^1.5.1
+  path_provider: ^2.0.2
 
 dev_dependencies:
   flutter_test:
